it's not only for products now
